// base/api/generated/models/RollenSystemRecht.ts
/* eslint-disable */
/**
 * dBildungs IAM
 * The dBildungs IAM server API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const RollenSystemRecht = {
    RollenVerwalten: 'ROLLEN_VERWALTEN',
    PersonenSofortLoeschen: 'PERSONEN_SOFORT_LOESCHEN',
    PersonenVerwalten: 'PERSONEN_VERWALTEN',
    LandesbediensteteSuchenUndHinzufuegen: 'LANDESBEDIENSTETE_SUCHEN_UND_HINZUFUEGEN',
    EingeschraenktNeueBenutzerErstellen: 'EINGESCHRAENKT_NEUE_BENUTZER_ERSTELLEN',
    SchulenVerwalten: 'SCHULEN_VERWALTEN',
    KlassenVerwalten: 'KLASSEN_VERWALTEN',
    SchultraegerVerwalten: 'SCHULTRAEGER_VERWALTEN',
    PersonSynchronisieren: 'PERSON_SYNCHRONISIEREN',
    CronDurchfuehren: 'CRON_DURCHFUEHREN',
    PersonenAnlegen: 'PERSONEN_ANLEGEN',
    ImportDurchfuehren: 'IMPORT_DURCHFUEHREN',
    PersonenLesen: 'PERSONEN_LESEN',
    BulkVerwalten: 'BULK_VERWALTEN',
    SchulportalVerwalten: 'SCHULPORTAL_VERWALTEN',
    HinweiseBearbeiten: 'HINWEISE_BEARBEITEN'
} as const;
export type RollenSystemRecht = typeof RollenSystemRecht[keyof typeof RollenSystemRecht];


export function RollenSystemRechtFromJSON(json: any): RollenSystemRecht {
    return RollenSystemRechtFromJSONTyped(json, false);
}

export function RollenSystemRechtFromJSONTyped(json: any, ignoreDiscriminator: boolean): RollenSystemRecht {
    return json as RollenSystemRecht;
}

export function RollenSystemRechtToJSON(value?: RollenSystemRecht | null): any {
    return value as any;
}